Notifications and messagges service updated

+++ src/modules/messages/services/messages.service.ts
@@ -1,4 +1,4 @@
-import { Injectable } from '@nestjs/common';
+import { HttpException, HttpStatus, Injectable } from '@nestjs/common';
 import { InjectRepository } from '@nestjs/typeorm';
 import { Repository } from 'typeorm';
 import { CreateMessageDto } from '../dto/create-message.dto';
@@ -15,23 +15,42 @@ export class MessagesService {
     private readonly messageRepository: Repository<MessageEntity>,
   ) {}
 
-
   // TO DO: Realizar validaciones
-
+  //throw new HttpException('Not found', HttpStatus.NOT_FOUND);
   // Function to create a new message
-  createMessage(createMessageDto: DeepPartial<CreateMessageDto>) { 
-    const message = this.messageRepository.create(createMessageDto)
-    return this.messageRepository.save(message)
- }
-
-  // Function to find all messages
-  findAllMessagesByChat(chatId : string): Promise<MessageEntity[]> { 
-    return this.messageRepository.find({ where: [ { chatId: chatId } ] } )
-   }
-
-  // Function to delete a message by ID
-  deleteMessage(messageId: string) { 
-    this.messageRepository.delete(messageId)
+  createMessage(createMessageDto: DeepPartial<CreateMessageDto>) {
+    try {
+      if (!createMessageDto) {
+        throw new HttpException('Not found', HttpStatus.NOT_FOUND);
+      }
+      const message = this.messageRepository.create(createMessageDto);
+      return this.messageRepository.save(message);
+    } catch (err) {
+      throw new HttpException('Bad Request', HttpStatus.BAD_REQUEST);
+    }
   }
 
+  // Function to find all messages
+  findAllMessagesByChat(chatId: string): Promise<MessageEntity[]> {
+    try {
+      if (!chatId) {
+        throw new HttpException('Not found', HttpStatus.NOT_FOUND);
+      }
+      return this.messageRepository.find({ where: [{ chatId: chatId }] });
+    } catch (err) {
+      throw new HttpException('Bad Request', HttpStatus.BAD_REQUEST);
+    }
+  }
+
+  // Function to delete a message by ID
+  deleteMessage(messageId: string) {
+    try {
+      if (!messageId) {
+        throw new HttpException('Not found', HttpStatus.NOT_FOUND);
+      }
+      this.messageRepository.delete(messageId);
+    } catch (err) {
+      throw new HttpException('Bad Request', HttpStatus.BAD_REQUEST);
+    }
+  }
 }
